change discard by right button to keybutton "escape"
# add class enum "WhatObjectAddition"
# create file and move functions of interaction buttons
