+++ main/mainwindow_connections.cpp
@@ -1,5 +1,3 @@
-#include <cmath>
-
 #include "mainwindow.h"
 // здесь описаны все соединения кнопок со слотами
 
@@ -14,6 +12,8 @@ gui::ObjectType MainWindow::GetObjType(CursorType cursor_type) {
       return gui::ObjectType::TrappyLines;
     case CursorType::HillCursor:
       return gui::ObjectType::Hills;
+    default:
+      return gui::ObjectType::Targets;
   }
 }
 
@@ -128,197 +128,6 @@ void MainWindow::on_pushButtonEditObjects_clicked() {
 void MainWindow::on_actionBeautify_triggered() {
   ui->plot->xAxis->setScaleRatio(ui->plot->yAxis, 1.0);
   ui->plot->replot();
-}
-
-void MainWindow::DisconnectObject(gui::ObjectType obj_type) {
-  switch (obj_type) {
-    case gui::ObjectType::TrappyCircles: {
-      disconnect(ui->plot, SIGNAL(mouseMove(QMouseEvent*)), this,
-                 SLOT(mouseMoveSetRadiusFromPlot(QMouseEvent*)));
-      disconnect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-                 SLOT(mousePressSetRadiusFromPlot(QMouseEvent*)));
-      disconnect(ui->plot, SIGNAL(mousePress(QMouseEvent*)), this,
-                 SLOT(mousePressDiscardTrappyCircle(QMouseEvent*)));
-      break;
-    }
-    case gui::ObjectType::TrappyLines: {
-      disconnect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-                 SLOT(mousePressSelectSecondTarget(QMouseEvent*)));
-      disconnect(ui->plot, SIGNAL(mousePress(QMouseEvent*)), this,
-                 SLOT(mousePressDiscardTrappyLine(QMouseEvent*)));
-      break;
-    }
-    case gui::ObjectType::Hills: {
-      disconnect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-                 SLOT(mousePressAddVertice(QMouseEvent*)));
-      disconnect(ui->plot, SIGNAL(mousePress(QMouseEvent*)), this,
-                 SLOT(mousePressDiscardHill(QMouseEvent*)));
-    }
-    default:
-      break;
-  }
-  connect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-          SLOT(mousePressObjectsButton(QMouseEvent*)));
-
-  ui->plot->setCursor(Qt::CrossCursor);
-  cursor_ = CursorType::DefaultCursor;
-}
-
-void MainWindow::mousePressDiscard(QMouseEvent* mouse_event) {
-  if (mouse_event->button() == Qt::RightButton) {
-    cursor_ = CursorType::DefaultCursor;
-    ui->plot->setCursor(Qt::CrossCursor);
-  }
-}
-
-void MainWindow::mousePressObjectsButton(QMouseEvent* mouse_event) {
-  double x = ui->plot->xAxis->pixelToCoord(mouse_event->pos().x());
-  double y = ui->plot->yAxis->pixelToCoord(mouse_event->pos().y());
-
-  switch (cursor_) {
-    case CursorType::TargetCursor: {
-      manager_->Add(new gui::Target(x, y));
-
-      ui->plot->setCursor(Qt::CrossCursor);
-      cursor_ = CursorType::DefaultCursor;
-      break;
-    }
-    case CursorType::TrCircleCursor: {
-      manager_->Add(new gui::TrappyCircle({x, y}, 0));
-      disconnect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-                 SLOT(mousePressObjectsButton(QMouseEvent*)));
-
-      connect(ui->plot, SIGNAL(mouseMove(QMouseEvent*)), this,
-              SLOT(mouseMoveSetRadiusFromPlot(QMouseEvent*)));
-      connect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-              SLOT(mousePressSetRadiusFromPlot(QMouseEvent*)));
-      connect(ui->plot, SIGNAL(mousePress(QMouseEvent*)), this,
-              SLOT(mousePressDiscardTrappyCircle(QMouseEvent*)));
-
-      ui->plot->setCursor(Qt::ClosedHandCursor);
-      break;
-    }
-    case CursorType::TrLineCursor: {
-      if (!ui->plot->selectedGraphs().empty()) {
-        for (const auto& t_ptr : manager_->GetTargetsPtrs()) {
-          if (t_ptr->GetGraphPtr() == ui->plot->selectedGraphs()[0]) {
-            manager_->Add(new gui::TrappyLine(t_ptr, t_ptr));
-            break;
-          }
-        }
-        disconnect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-                   SLOT(mousePressObjectsButton(QMouseEvent*)));
-
-        connect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-                SLOT(mousePressSelectSecondTarget(QMouseEvent*)));
-        connect(ui->plot, SIGNAL(mousePress(QMouseEvent*)), this,
-                SLOT(mousePressDiscardTrappyLine(QMouseEvent*)));
-      }
-      break;
-    }
-    case CursorType::HillCursor: {
-      manager_->Add(new gui::Hill{{x, y}, {x, y}});
-
-      disconnect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-                 SLOT(mousePressObjectsButton(QMouseEvent*)));
-
-      connect(ui->plot, SIGNAL(mouseDoubleClick(QMouseEvent*)), this,
-              SLOT(mousePressAddVertice(QMouseEvent*)));
-      connect(ui->plot, SIGNAL(mousePress(QMouseEvent*)), this,
-              SLOT(mousePressDiscardHill(QMouseEvent*)));
-      break;
-    }
-    default:
-      break;
-  }
-  area_->Redraw();
-  t_connection_->UpdateTables();
-}
-
-void MainWindow::mouseMoveSetRadiusFromPlot(QMouseEvent* mouse_event) {
-  size_t last = manager_->GetTrappyCircles().size() - 1;
-
-  // Расстояние от центра до курсора (по оси x)
-  double x = abs(ui->plot->xAxis->pixelToCoord(mouse_event->pos().x()) -
-                 manager_->GetTrappyCircles()[last].GetCenter().x);
-
-  // Расстояние от центра до курсора (по оси y)
-  double y = abs(ui->plot->yAxis->pixelToCoord(mouse_event->pos().y()) -
-                 manager_->GetTrappyCircles()[last].GetCenter().y);
-
-  // Значение радиуса
-  double r = pow(pow(x, 2) + pow(y, 2), 0.5);
-
-  manager_->GetTrappyCirclesPtrs()[last]->SetRadius(r);
-  area_->Redraw();
-  t_connection_->UpdateTables();
-}
-
-void MainWindow::mousePressDiscardTrappyCircle(QMouseEvent* mouse_event) {
-  if (mouse_event->button() == Qt::RightButton) {
-    size_t last = manager_->GetTrappyCircles().size() - 1;
-    manager_->Remove(gui::ObjectType::TrappyCircles, last);
-
-    DisconnectObject(gui::ObjectType::TrappyCircles);
-    area_->Redraw();
-    t_connection_->UpdateTables();
-  }
-}
-
-void MainWindow::mousePressSelectSecondTarget(QMouseEvent* mouse_event) {
-  if (!ui->plot->selectedGraphs().empty()) {
-    for (const auto& t_ptr : manager_->GetTargetsPtrs()) {
-      if (t_ptr->GetGraphPtr() == ui->plot->selectedGraphs()[0]) {
-        size_t last = manager_->GetTrappyLines().size() - 1;
-        if (manager_->GetTrappyLinesPtrs()[last]->GetTargetsPtrs().second !=
-            t_ptr) {
-          manager_->GetTrappyLinesPtrs()[last]->SetSecondTarget(t_ptr);
-
-          DisconnectObject(gui::ObjectType::TrappyLines);
-          area_->Redraw();
-          t_connection_->UpdateTables();
-          break;
-        }
-      }
-    }
-  }
-}
-
-void MainWindow::mousePressDiscardTrappyLine(QMouseEvent* mouse_event) {
-  if (mouse_event->button() == Qt::RightButton) {
-    size_t last = manager_->GetTrappyLines().size() - 1;
-    manager_->Remove(gui::ObjectType::TrappyLines, last);
-
-    DisconnectObject(gui::ObjectType::TrappyLines);
-    area_->Redraw();
-    t_connection_->UpdateTables();
-  }
-}
-
-void MainWindow::mousePressAddVertice(QMouseEvent* mouse_event) {
-  double x = ui->plot->xAxis->pixelToCoord(mouse_event->pos().x());
-  double y = ui->plot->yAxis->pixelToCoord(mouse_event->pos().y());
-  size_t last = manager_->GetHills().size() - 1;
-
-  if (abs(manager_->GetHills()[last].GetPoints()[0].x - x) < 0.1 &&
-      abs(manager_->GetHills()[last].GetPoints()[0].y - y) < 0.1) {
-    DisconnectObject(gui::ObjectType::Hills);
-  } else
-    manager_->GetHillsPtrs()[last]->AddVertice({x, y});
-
-  area_->Redraw();
-  t_connection_->UpdateTables();
-}
-
-void MainWindow::mousePressDiscardHill(QMouseEvent* mouse_event) {
-  if (mouse_event->button() == Qt::RightButton) {
-    size_t last = manager_->GetHills().size() - 1;
-    manager_->Remove(gui::ObjectType::Hills, last);
-
-    DisconnectObject(gui::ObjectType::Hills);
-    area_->Redraw();
-    t_connection_->UpdateTables();
-  }
 }
 
 // MARK: Json connections
